+++ react-team-project/package.json
@@ -30,7 +30,7 @@
     "react-router-dom": "^6.22.3",
     "react-scroll-parallax": "^3.4.5",
     "react-transition-group": "^4.4.5",
-    "rsuite": "^5.56.0",
+    "rsuite": "^5.57.0",
     "swr": "^2.2.4"
   },
   "devDependencies": {
